Turn `Iter`’s impls into a macro for any newtype.
This will allow us to easily define iterator newtypes that forward all
the methods useful to forward, instead of forwarding only some methods
or lazily exposing the underlying iterator.
